LPS-164331 Move ComponentDescriptor, SoyComponentRendererImpl, Helper, and boostrap.js.tpl to frontend.taglib.soy.internal.template
