+++ modules/apps/frontend-taglib/frontend-taglib-soy/src/main/java/com/liferay/frontend/taglib/soy/internal/template/SoyComponentRendererImpl.java
@@ -12,14 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.template.soy.renderer.internal;
+package com.liferay.frontend.taglib.soy.internal.template;
 
 import com.liferay.frontend.js.module.launcher.JSModuleLauncher;
 import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.template.soy.renderer.ComponentDescriptor;
 import com.liferay.portal.template.soy.renderer.SoyComponentRenderer;
-import com.liferay.portal.template.soy.renderer.internal.helper.SoyComponentRendererHelper;
 
 import java.io.IOException;
 import java.io.Writer;
