+++ modules/apps/frontend-taglib/frontend-taglib-soy/src/main/resources/com/liferay/frontend/taglib/soy/internal/template/dependencies/bootstrap.js.tpl
